trading_system/requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
websocket-client>=1.3.0
web3>=6.0.0
python-dotenv>=0.19.0
psutil>=5.8.0
pytest>=7.0.0
aiohttp>=3.8.0
scikit-learn>=1.0.0
